+++ app/src/main/res/layout/detalle_album.xml
@@ -13,7 +13,7 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_marginTop="10dp"
-        android:layout_marginBottom="10dp"
+        android:layout_marginBottom="12dp"
         android:background="@android:color/transparent"
         android:backgroundTint="@android:color/transparent">
