+++ gradle/wrapper/gradle-wrapper.properties
@@ -1,4 +1,4 @@
-#Sat Nov 06 14:26:21 COT 2021
+#Sat Nov 06 19:28:40 COT 2021
 distributionBase=GRADLE_USER_HOME
 distributionUrl=https\://services.gradle.org/distributions/gradle-7.0.2-bin.zip
 distributionPath=wrapper/dists
